Factor out util.ts.
