+++ exporter/export.ts
@@ -1,36 +1,12 @@
-import * as fs from "fs";
 import * as path from "path";
-import * as child_process from "child_process";
 import { syncVersion } from "./sync-version";
-import { GODOT_PATH, ROOT_DIR } from "./paths";
+import { GODOT_PATH } from "./paths";
+import { ensureDirIsEmptySync, runSync } from "./util";
 
 // TODO: support publishing to itch via e.g.:
 //   butler push dist toolness/cat-in-the-coop:html --userversion 0.0.1
 //   butler push cat-in-the-coop-osx.zip toolness/cat-in-the-coop:osx-beta --userversion 0.0.1
 //   butler push dist-windows toolness/cat-in-the-coop:windows-beta --userversion 0.0.1
-
-function runSync(path: string, args: string[]) {
-  const result = child_process.spawnSync(
-    path, args, {
-      cwd: ROOT_DIR,
-      stdio: 'inherit',
-    }
-  );
-  if (result.status || result.status === null) {
-    console.log("Subprocess failed, exiting.");
-    process.exit(1);
-  }
-}
-
-function ensureDirIsEmptySync(dirName: string) {
-  if (!fs.existsSync(dirName)) {
-    fs.mkdirSync(dirName);
-  }
-  
-  for (let filename of fs.readdirSync(dirName)) {
-    fs.unlinkSync(path.join(dirName, filename));
-  }
-}
 
 interface Exporter {
   name: string;
